refactor(SliceSumCat): move fuse_slice_sum_cat to structure patterns * fix: make fuse_slice_sum_cat a structure pattern * fix: separate out the constraint_fn and default replacement module for slice_sum_cat module
